refactor mostly stable (81%)
