+++ src/notvox/__init__.py
@@ -1,3 +1,4 @@
+# client/src/notvox/__init__.py
 """
 NotVox - Network Spotify Control System
 """
@@ -10,3 +11,15 @@ __email__ = "espadonn@outlook.com"
 from .cli import cli
 
 __all__ = ["cli"]
+
+
+# client/src/notvox/commands/__init__.py
+"""NotVox command modules"""
+
+# This can be empty - commands are imported directly in cli.py
+
+
+# client/src/notvox/utils/__init__.py
+"""NotVox utility modules"""
+
+# This can be empty - utils are imported where needed
